Clean up `mill.eval` even more (#2542)
1. `Terminal` goes in its own file
2. Split out `EvaluatorCore` (which contains only evaluation-related
stuff) from `EvaluatorImpl` (which contains some user-facing helpers)
3. Split out `GroupEvaluator` from `EvaluatorCore`
4. Split out `Plan`, `ThreadNumberer`, `JsonArrayLogger`
5. Unify implementation of `mill-profile.json` and
`mill-par-profile.json` (renamed `mill-chrome-profile.json`), so they
both perform streaming as-evaluation-progresses logging in a thread-safe
manner (should fix https://github.com/com-lihaoyi/mill/issues/2540)
6. Add `dependencies` to `mill-profile.json`, making it more useful
since you can now trace up and down the build graph without having to
run `mill inspect` (which would wipe out the `mill-profile.json` you
were just looking at!)
